TestClassBuilderHelper refactoring
