+++ src/main/kotlin/org/jetbrains/research/testspark/helpers/LLMHelper.kt
@@ -262,14 +262,14 @@ object LLMHelper {
         }
 
         val jUnitVersion = project.getService(LLMSettingsService::class.java).state.junitVersion
-        val testBodyPrinter = TestBodyPrinterFactory.createTestBodyPrinter(language)
+        val testBodyPrinter = TestBodyPrinterFactory.create(language)
         val testSuiteParser = TestSuiteParserFactory.createJUnitTestSuiteParser(
             jUnitVersion,
             language,
             testBodyPrinter,
         )
 
-        val testsAssembler = TestsAssemblerFactory.createTestsAssembler(
+        val testsAssembler = TestsAssemblerFactory.create(
             indicator,
             testGenerationOutput,
             testSuiteParser,
