+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/test/JUnitTestSuitePresenter.kt
@@ -5,8 +5,7 @@ import org.jetbrains.research.testspark.core.data.TestGenerationData
 import org.jetbrains.research.testspark.core.generation.llm.getClassWithTestCaseName
 import org.jetbrains.research.testspark.core.test.SupportedLanguage
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
-import org.jetbrains.research.testspark.helpers.java.JavaClassBuilderHelper
-import org.jetbrains.research.testspark.helpers.kotlin.KotlinClassBuilderHelper
+import org.jetbrains.research.testspark.tools.TestClassCodeGeneratorFactory
 
 class JUnitTestSuitePresenter(
     private val project: Project,
@@ -37,29 +36,16 @@ class JUnitTestSuitePresenter(
             // Add each test
             testCases.forEach { testCase -> testBody += "$testCase\n" }
 
-            when (language) {
-                SupportedLanguage.Java -> JavaClassBuilderHelper.generateCode(
-                    project,
-                    testFileName,
-                    testBody,
-                    imports,
-                    packageName,
-                    runWith,
-                    otherInfo,
-                    generatedTestsData,
-                )
-
-                SupportedLanguage.Kotlin -> KotlinClassBuilderHelper.generateCode(
-                    project,
-                    testFileName,
-                    testBody,
-                    imports,
-                    packageName,
-                    runWith,
-                    otherInfo,
-                    generatedTestsData,
-                )
-            }
+            TestClassCodeGeneratorFactory.create(language).generateCode(
+                project,
+                testFileName,
+                testBody,
+                imports,
+                packageName,
+                runWith,
+                otherInfo,
+                generatedTestsData,
+            )
         }
     }
 
@@ -73,29 +59,16 @@ class JUnitTestSuitePresenter(
         testCaseIndex: Int,
     ): String =
         testSuite.run {
-            when (language) {
-                SupportedLanguage.Java -> JavaClassBuilderHelper.generateCode(
-                    project,
-                    getClassWithTestCaseName(testCases[testCaseIndex].name),
-                    testCases[testCaseIndex].toStringWithoutExpectedException() + "\n",
-                    imports,
-                    packageName,
-                    runWith,
-                    otherInfo,
-                    generatedTestsData,
-                )
-
-                SupportedLanguage.Kotlin -> KotlinClassBuilderHelper.generateCode(
-                    project,
-                    getClassWithTestCaseName(testCases[testCaseIndex].name),
-                    testCases[testCaseIndex].toStringWithoutExpectedException() + "\n",
-                    imports,
-                    packageName,
-                    runWith,
-                    otherInfo,
-                    generatedTestsData,
-                )
-            }
+            TestClassCodeGeneratorFactory.create(language).generateCode(
+                project,
+                getClassWithTestCaseName(testCases[testCaseIndex].name),
+                testCases[testCaseIndex].toStringWithoutExpectedException() + "\n",
+                imports,
+                packageName,
+                runWith,
+                otherInfo,
+                generatedTestsData,
+            )
         }
 
     /**
@@ -110,30 +83,16 @@ class JUnitTestSuitePresenter(
             // Add each test (exclude expected exception)
             testCases.forEach { testCase -> testBody += "${testCase.toStringWithoutExpectedException()}\n" }
 
-            when (language) {
-                SupportedLanguage.Java ->
-                    JavaClassBuilderHelper.generateCode(
-                        project,
-                        testFileName,
-                        testBody,
-                        imports,
-                        packageName,
-                        runWith,
-                        otherInfo,
-                        generatedTestsData,
-                    )
-
-                SupportedLanguage.Kotlin -> KotlinClassBuilderHelper.generateCode(
-                    project,
-                    testFileName,
-                    testBody,
-                    imports,
-                    packageName,
-                    runWith,
-                    otherInfo,
-                    generatedTestsData,
-                )
-            }
+            TestClassCodeGeneratorFactory.create(language).generateCode(
+                project,
+                testFileName,
+                testBody,
+                imports,
+                packageName,
+                runWith,
+                otherInfo,
+                generatedTestsData,
+            )
         }
     }
